+++ packages/auth-zkp-iden3/src/helpers/file-helpers.ts
@@ -1,0 +1,5 @@
+export async function readBytesFile(path: string) {
+  const res = await fetch(path)
+  const buff = await res.arrayBuffer()
+  return new Uint8Array(buff)
+}