+++ packages/auth-zkp-iden3/src/helpers/index.ts
@@ -1,0 +1,2 @@
+export * from './file-helpers'
+export * from './state-v2-helpers'