+++ packages/auth-zkp-iden3/src/types/index.ts
@@ -1,0 +1,2 @@
+export * from './auth-zkp-types'
+export * from './contracts'